+++ src/models/regional/containers.ts
@@ -43,4 +43,24 @@ export interface ContainerSize {
   /** The pricing grid for this container size */
   pricings: PricingGrid;
 }
+
+export interface ContainerProcess {
+  /** Unique universal identifier */
+  id: string;
+  /** Process type, as defined in the Procfile */
+  type: string;
+  /** index of the container for this type */
+  type_index: number;
+  /** type+index of the container for this type */
+  label: number;
+  /** Creation date of this container */
+  created_at: string;
+  /** Deletion date of this container */
+  deleted_at: null;
+  /** State of this container */
+  state: string;
+  /** The command ran by this container */
+  command: string;
+  /** The container size */
+  container_size: ContainerSize;
 }